refact: refact stracture
