+++ src/circuit.rs
