+++ src/lambda_residues.rs
@@ -64,8 +64,6 @@ impl LambdaResidues {
         assert_ne!(w.pow(cofactor_cubic.to_u64_digits()), ark_bn254::Fq12::ONE);
         assert_eq!(w.pow(h.to_u64_digits()), ark_bn254::Fq12::ONE);
 
-        println!("w :{:?}", w.to_string());
-
         // just two option, w and w^2, since w^3 must be cubic residue, leading f*w^3 must not be cubic residue
         let mut wi = w;
         if (f * wi).pow(cofactor_cubic.to_u64_digits()) != ark_bn254::Fq12::ONE {
@@ -97,17 +95,7 @@ impl LambdaResidues {
         // d-th (cubic) root, say c
         let c = Self::tonelli_shanks_cubic(f3, w, s, t, k);
         assert_ne!(c, ark_bn254::Fq12::ONE);
-        let left = c.pow(params::LAMBDA.clone().to_u64_digits());
-        let right = f * wi;
-
-        println!("c :{:?}", c.to_string());
-        println!("lambda :{:?}", params::LAMBDA.clone().to_string());
-        println!("c^lambda :{:?}", left.to_string());
-        println!("\n f :{:?}", f.to_string());
-        println!("wi :{:?}", wi.to_string());
-        println!("f*wi :{:?}", right.to_string());
-
-        assert_eq!(left, right);
+        assert_eq!(c.pow(params::LAMBDA.clone().to_u64_digits()), f * wi);
 
         Self { c, wi }
     }
@@ -170,7 +158,7 @@ mod test {
     use ark_std::UniformRand;
     use num_bigint::BigUint;
 
-    use crate::{dev, params};
+    use crate::params;
     use ark_bn254::{Bn254, G1Affine, G2Affine};
     use ark_ec::pairing::Pairing;
     use ark_ec::{AffineRepr, CurveGroup};
@@ -254,36 +242,5 @@ mod test {
         // d-th (cubic) root, say c
         let c = LambdaResidues::tonelli_shanks_cubic(f3, w, s, t, k);
         assert_eq!(c.pow(params::LAMBDA.deref().to_u64_digits()), f * wi);
-    }
-
-    #[test]
-    #[ignore]
-    fn test_finding_c_witn_ark_miller_loop() {
-        let Q = dev::g2
-            .mul_bigint(BigUint::from_i8(3).unwrap().to_u64_digits())
-            .into_affine();
-        let P = dev::g1
-            .mul_bigint(BigUint::from_i8(4).unwrap().to_u64_digits())
-            .into_affine();
-
-        println!("P:{:?}", P);
-        println!("\n Q:{:?}", Q);
-        let actual = Bn254::miller_loop(
-            <G1Affine as Into<<Bn254 as Pairing>::G1Prepared>>::into(P),
-            <G2Affine as Into<<Bn254 as Pairing>::G2Prepared>>::into(Q),
-        );
-
-        let residues = LambdaResidues::finding_c(actual.0);
-    }
-
-    #[test]
-    fn test_finding_c_witn_native_miller_loop() {
-        let Q = dev::g2.mul_bigint(BigUint::from_i8(3).unwrap().to_u64_digits());
-        let P = dev::g1.mul_bigint(BigUint::from_i8(4).unwrap().to_u64_digits());
-
-        println!("P:{:?}", P);
-        println!("\n Q:{:?}", Q);
-        // let actual = NativeMillerLoop::miller_loop(P, Q);
-        // let residues = LambdaResidues::finding_c(actual);
     }
 }
