+++ src/lib.rs
@@ -1,12 +1,8 @@
 #[cfg(test)]
-pub mod dev;
-// pub mod ell_coeffs;
-#[cfg(test)]
-pub mod groth16;
+pub mod circuit;
 pub mod groth16_verifier;
 pub mod lambda_residues;
-pub mod miller_lines;
-pub mod optimal_ate;
-pub mod pairing_verify;
+pub mod pairing_verifier;
 pub mod params;
-pub mod utils;
+mod poc;
+mod utils;
