+++ src/pairing_verifier.rs
@@ -1,0 +1,153 @@
+use crate::params;
+use ark_bn254::{fq::Fq, Bn254, Fq12, Fq2, Fq6, G1Affine, G2Affine, G2Projective};
+use ark_ec::bn::g2::{mul_by_char, G2HomProjective};
+use ark_ec::bn::{BnConfig, G2Prepared};
+use ark_ec::AffineRepr;
+use ark_ff::Field;
+use ark_std::UniformRand;
+use num_bigint::BigUint;
+use num_traits::One;
+use rand::SeedableRng;
+use std::ops::{Mul, Neg};
+
+// To verify e(P1,Q1)*e(P2,Q2)*e(P3,Q3)*e(P4,Q4)=1
+//
+// Here is only support to verify groth16's pairing, which (Q1,Q2,Q3) are fixed, Q4 is non-fixed.
+//
+// params:
+//  @eval_points: [P1,P2,P3]. which has fixed {Q1,Q2,Q3}
+//  @P4: P4
+//  @Q4: Q4
+//  @lines: []precompute miller lines for Qi. Only support fixed Qi.
+//  @c: c^lambda = f*w^i
+//  @c_inv: inverse of c
+//
+// verify c^lambda = f * wi, namely c_inv^lambda * f * wi = 1
+pub fn quad_miller_loop_with_c_wi(
+    eval_points: Vec<G1Affine>,
+    P4: G1Affine,
+    Q4: G2Affine,
+    // lines: &[Vec<(Fq2, Fq2)>],
+    lines: &Vec<G2Prepared<ark_bn254::Config>>,
+    c: Fq12,
+    c_inv: Fq12,
+    wi: Fq12,
+    // TODO: What's B in stack
+) -> Fq12 {
+    assert_eq!(eval_points.len(), 3, "Should contains 4 G1Affine: P1,P2,P3");
+    assert_eq!(lines.len(), 3, "Only precompute lines for Q1,Q2,Q3");
+    assert_eq!(c * c_inv, Fq12::ONE, "Check if c·c^−1 = 1");
+
+    let mut T4 = G2HomProjective::<ark_bn254::Config> {
+        x: Q4.x,
+        y: Q4.y,
+        z: ark_bn254::Fq2::one(),
+    };
+
+    // constants
+    let two_inv = ark_bn254::Fq::one().double().inverse().unwrap();
+
+    // 1. f = c_inv
+    let mut f = c_inv;
+
+    let mut lines_iters = lines
+        .iter()
+        .map(|item| item.ell_coeffs.iter())
+        .collect::<Vec<_>>();
+
+    // 2. miller loop part, 6x + 2
+    for i in (1..ark_bn254::Config::ATE_LOOP_COUNT.len()).rev() {
+        let bit = ark_bn254::Config::ATE_LOOP_COUNT[i - 1];
+
+        // 2.1 double: f = f * f
+        f = f.square();
+
+        // 2.2 mul c
+        //  f = f * c_inv, if digit == 1
+        //  f = f * c, if digit == -1
+        f = if 1 == bit {
+            f * c_inv
+        } else if bit == -1 {
+            f * c
+        } else if bit == 0 {
+            f
+        } else {
+            panic!("bit is not in (-1,1), bit={bit}");
+        };
+
+        // 2.3 accumulate double lines (fixed and non-fixed)
+        // 2.3.1(fixed) f = f * double_line_Q(P). fixed points: P1, P2, P3
+        for (line_i, pi) in lines_iters.iter_mut().zip(eval_points.iter()) {
+            let line_i_0 = line_i.next().unwrap();
+            Bn254::ell(&mut f, line_i_0, pi);
+        }
+
+        // 2.3.2(non-fixed) double line with T4 (projective coordinates)
+        let double_line = T4.double_in_place(&two_inv); // TODO: check if the param is 1/2
+
+        // 2.3.3(non-fixed) evaluation double_line. non-fixed points: P4
+        Bn254::ell(&mut f, &double_line, &P4);
+
+        if bit == 1 || bit == -1 {
+            // 2.4 accumulate add lines (fixed and non-fixed)
+            // 2.4.1(fixed) f = f * add_line_eval. fixed points: P1, P2, P3
+            for (line_i, pi) in lines_iters.iter_mut().zip(eval_points.iter()) {
+                let line_i_1 = line_i.next().unwrap();
+                Bn254::ell(&mut f, line_i_1, pi);
+            }
+            // 2.4.2(non-fixed) double line with T4 (projective coordinates)
+            let add_line = if bit == 1 {
+                T4.add_in_place(&Q4)
+            } else {
+                // }else if bit == -1 {
+                T4.add_in_place(&Q4.neg())
+            };
+
+            // 2.4.3(non-fixed) evaluation double_line. non-fixed points: P4
+            Bn254::ell(&mut f, &add_line, &P4);
+        }
+    }
+
+    // 3. f = f * c_inv^p * c^{p^2}
+    f = f
+        * c_inv.pow(params::MODULUS.to_u64_digits())
+        * c.pow(params::MODULUS.pow(2).to_u64_digits());
+
+    // 4. f = f * wi . scale f
+    f = f * wi;
+
+    // 5 add lines (fixed and non-fixed)
+    // 5.1(fixed) f = f * add_line_eval. fixed points: P1, P2, P3
+    for (line_i, pi) in lines_iters.iter_mut().zip(eval_points.iter()) {
+        let line_i_1 = line_i.next().unwrap();
+        Bn254::ell(&mut f, line_i_1, pi);
+    }
+    // 5.2 one-time frobenius map to compute phi_Q
+    //     compute phi(Q) with Q4
+    let phi_Q = mul_by_char::<ark_bn254::Config>(Q4.clone());
+
+    let add_line = T4.add_in_place(&phi_Q);
+
+    // 5.4(non-fixed) evaluation add_lin. non-fixed points: P4
+    Bn254::ell(&mut f, &add_line, &P4);
+
+    // 6. add lines (fixed and non-fixed)
+    // 6.1(fixed) f = f * add_line_eval. fixed points: P1, P2, P3
+    for (line_i, pi) in lines_iters.iter_mut().zip(eval_points.iter()) {
+        // TODO: where is f?? and where is double line?
+        let line_i_1 = line_i.next().unwrap();
+        Bn254::ell(&mut f, line_i_1, pi);
+    }
+    // 6.2 two-time frobenius map to compute phi_Q
+    //     compute phi_Q_2 with phi_Q
+    // mul_by_char: used to q's frob...map.
+    let mut phi_Q_2 = mul_by_char::<ark_bn254::Config>(phi_Q.clone());
+    phi_Q_2.y = phi_Q_2.y.neg();
+    let add_line = T4.add_in_place(&phi_Q_2);
+
+    // 6.3(non-fixed) evaluation add_lin. non-fixed points: P4
+    Bn254::ell(&mut f, &add_line, &P4);
+
+    // return final_f
+    f
+}
